pong_project_py27/requirements.txt
appdirs==1.4.3
Django==1.6
easy-thumbnails==2.4.1
olefile==0.44
packaging==16.8
Pillow==4.1.1
pygame==1.9.3
pyparsing==2.2.0
SimpleGUICS2Pygame==1.9.0
SimpleGUITk==1.1.3
six==1.10.0
